% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Make an animated gif of landscape over time}
\usage{
animate(lands, stepsBy = 1, steps = NULL, fps = 5, gifName = NULL,
  rangeLimitOccup = 0.75, width = NULL, height = NULL)
}
\arguments{
\item{lands}{the object output of the \code{\link{run_model}} function}

\item{stepsBy}{integer number, increment value from 1 to length of steps}

\item{steps}{vector, specific steps to be animated}

\item{fps}{frames per second}

\item{gifName}{string with name of the file output}

\item{rangeLimitOccup}{numeric between 0 and 1 to define the minimum row occupancy of a state in the landscape. It will add a line in the plot for the boreal trailing edge and the temperate leading edge. See \code{\link{run_model}} for more details}

\item{width}{numeric, width of gif image}

\item{height}{numeric, height of gif image}
}
\value{
an animated gif file
}
\description{
This function creates an animated gif with every landscape year configuration
}
\examples{
\dontrun{
animate(lands, stepsBy = 1, fps = 5, gifName = 'myGif', rangeLimitOccup = 0.75)
}
}
