% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landscape.R
\name{create_virtual_landscape}
\alias{create_virtual_landscape}
\title{Create virtual landscape}
\usage{
create_virtual_landscape(climRange = c(-2.61, 5.07), cellSize = 0.8)
}
\arguments{
\item{climRange}{vector, annual mean temperature with the respective North and South limits. Values must be between -5.3 and 12.2 to respect the parameterization boundary.}

\item{cellSize}{numeric, size of the cell of the landscape in Km. Recommended cell size is between 0.3 and 5 km.}
}
\value{
a list with the (i) initial landscape, (ii) scaled temperature gradient, (iii) landscape dimensions, (iv) position and neighbor are internal objects to run the model in parallel.
}
\description{
This function creates an initial virtual landscape to run the simulations
}
\examples{
initLand = create_virtual_landscape(climRange = c(-2.61, 5.07), cellSize = 4)
}
