% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_limit.R
\name{range_limit}
\alias{range_limit}
\title{Calculate state range limit}
\usage{
range_limit(land, nRow, nCol, occup)
}
\arguments{
\item{land}{vector, one element of the \code{\link{run_model}} output}

\item{nRow}{numeric, number of rows of the landscape. Value is found in the output list from the \code{\link{run_model}} function}

\item{nCol}{numeric, number of columns of the landscape. Value is found in the output list from the \code{\link{run_model}} function}

\item{occup}{numeric between 0 and 1. The value determines the minimum occupancy a row of the landscape must be occupied by a specific forest state to be considered part of the state range}
}
\description{
This function calculates the south range limit of boreal and the north range limit of temperate states in a specific landscape configuration
}
\examples{
\dontrun{
lands <- run_model(steps = 10, initLand)
range_limit(lands[['land_TX']], lands[['nRow']], lands[['nCol']], occup = .7)
}
}
